task: Update docker compose network configuration
I've removed the unnecessary repetition of the ledger network
configuration from the various containers.  They will all be added to
the default network... by default.  I've simply added a name to it so
that it can be easily referenced from other local docker network
contexts (i.e. publishers).  A companion PR outlining some of this in
publishers is imminent.