Improvement/publishers network (#1055)
* docs: Update README to reflect current docker nomenclature

* task: Update docker compose network configuration

I've removed the unnecessary repetition of the ledger network
configuration from the various containers.  They will all be added to
the default network... by default.  I've simply added a name to it so
that it can be easily referenced from other local docker network
contexts (i.e. publishers).  A companion PR outlining some of this in
publishers is imminent.

* task: Add to docker file for simplicity

* improvement: Makefile

* docs: Update README for running individual tests

* task: Update graphana docker compose port

* Got Zookeeper and Kafka to boot up properly without SSL

* Simplify the docker compose setup

* Remove SSL in dev, simplify compose file

* Remove dev compose from package.json

* Update README

Co-authored-by: Tyler Smart <tsmart@brave.com>