Re-enable SSL in dev
The dev/test certs had expired, so I regenerated them using the
certs script. This has to be run on the kafka container so that the
same java version that creates the certs will read the certs. I also
had to change the CN name to match the DNS name specified in the
docker-compose, namely kafka1. I also re-added the decrpytion password
to the SSL options.

This is a fairly brittle system and is difficult to diagnose issues when
they go wrong, but allows us to run SSL in dev and keep the codebase
simplified.